+++ src/main/java/com/playko/projectManagement/application/dto/request/team/AddUserToTeamRequestDto.java
@@ -1,0 +1,15 @@
+package com.playko.projectManagement.application.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class AddUserToTeamRequestDto {
+    private Long teamId;
+    private String emailUser;
+}
